pyproject.toml
[project]
name = "cookmarks"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "gunicorn>=22.0.0",
    "whitenoise>=6.7.0",
    "lxml>=6.0.2",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "requests>=2.32.5",
    "django-q2>=1.6.1",
    "tiktoken>=0.12.0",
    "google-genai>=1.3.0",
]
